+++ src/main/java/tornadofx/FX.kt
@@ -356,12 +356,7 @@ fun varargParamsToMap(params: Array<out Pair<String, Any?>>): Map<*, Any?>? {
 
 @Suppress("UNCHECKED_CAST")
 fun <T : Component> find(type: KClass<T>, scope: Scope = DefaultScope, params: Map<*, Any?>? = null): T {
-    val useScope = if (InstanceScoped::class.java.isAssignableFrom(type.java)) {
-        // InstanceScoped types must keep their scope if already instantiated, else create a new scope
-        if (Injectable::class.java.isAssignableFrom(type.java) && FX.getComponents(scope).containsKey(type as KClass<out Injectable>)) scope else Scope()
-    } else {
-        FX.fixedScopes[type] ?: scope
-    }
+    val useScope = FX.fixedScopes[type] ?: scope
     inheritScopeHolder.set(useScope)
     val stringKeyedMap = HashMap<String, Any?>()
     params?.forEach {