+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 - removeWhen/visibleWhen/enableWhen/disableWhen etc functions now also take an observable instead of a function that returns an observable.
 - The `label` builder is now capable of taking a graphic node `label("some text", graphic)`
+- ComboBoxBase.required() validator
+- SmartResize.POLICY can now be installed by calling `smartResize()` on any `TableView`
+- SmartResize will automatically resize if the itemsProperty of the TableView changes value
 
 ### Fixed
 
@@ -16,7 +19,8 @@
 
 - CSS warning should not be issued in OSGi environment, since bundle activator installs CSS URL Handler
 - All shape builders accepts `Number` instead of `Double` so you can write `circle(10, 10, 5)` instead of `circle(10.0, 10.0, 5.0)`
-- Deprecated `menuitem` builders in favor of `item` builders, which work the same way as other builders with respect to action (IDEA provides quick fix)
+- ComboBox.validator moved to ComboBoxBase.validator to support ColorPicker and DatePicker as well
+- Removed InstanceScoped and removed it from Wizard. It was not needed.
 
 ## [1.7.4] - 2017-04-28